+++ packages/client/src/index.ts
@@ -1,6 +1,5 @@
 /* istanbul ignore file */
 
-export { default as HWTransport } from "@ledgerhq/hw-transport";
 export * from "@ledgerhq/wallet-api-core";
 export * from "./TransportWalletAPI";
 export * from "./WalletAPIClient";